pyproject.toml
[tool.poetry]
name = "terrarium-capp"
version = "0.2.0"
description = "Adaptive Computer Aided Process Planner"
authors = ["Justin Vrana <justin.vrana@gmail.com>"]
homepage = "https://www.github.com/jvrana/Terrarium"
repository = "https://www.github.com/jvrana/Terrarium"
documentation = "https://www.github.com/jvrana/Terrarium"
readme = 'README.md'
packages = [
    { include = "terrarium", from = "." },
]
classifies = [
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Life",
    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7"
]

[tool.poetry.dependencies]
python = "^3.7"
dill = "^0.2.9"
webcolors = "^1.9"
colorama = "^0.4.1"
termcolor = "^1.1"
arrow = "^0.14.2"
pydent = "=0.1.5a2"
loggable-jdv = "^0.1.2"

[tool.poetry.dev-dependencies]
black = { version = "^19.3b0", python = "^3.6" }
pytest = "^4.6"
pre-commit = "^1.17"
vcrpy = "^2.0"
pytest-cov = "^2.7"
toml = "^0.10.0"
keats = "^0.2.25"
tox = "^3.13"
tox-conda = "^0.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
